+++ src/Controller/Utils/StoreTrait.php
@@ -32,6 +32,7 @@ use AppBundle\Service\InvitationManager;
 use AppBundle\Sylius\Order\OrderFactory;
 use AppBundle\Sylius\Order\OrderInterface;
 use Carbon\Carbon;
+use Cocur\Slugify\SlugifyInterface;
 use Doctrine\ORM\QueryBuilder;
 use Doctrine\ORM\EntityManagerInterface;
 use Hashids\Hashids;
@@ -48,6 +49,7 @@ use Sylius\Bundle\OrderBundle\NumberAssigner\OrderNumberAssignerInterface;
 use Symfony\Component\Form\FormError;
 use Symfony\Component\Form\FormInterface;
 use Symfony\Component\Form\Extension\Core\Type\EmailType;
+use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\ResponseHeaderBag;
@@ -332,7 +334,7 @@ trait StoreTrait
 
                 $order = $delivery->getOrder();
                 $this->handleBookmark($orderManager, $form, $order);
-                $this->handleNewSubscription($pricingManager, $logger, $store, $form, $delivery, $order, new UseArbitraryPrice($arbitraryPrice));
+                $this->handleNewRecurrenceRule($pricingManager, $logger, $store, $form, $delivery, $order, new UseArbitraryPrice($arbitraryPrice));
 
                 $entityManager->flush();
 
@@ -355,7 +357,7 @@ trait StoreTrait
                     // because an auto increment is needed to generate a number
                     $orderManager->onDemand($order);
 
-                    $this->handleNewSubscription($pricingManager, $logger, $store, $form, $delivery, $order);
+                    $this->handleNewRecurrenceRule($pricingManager, $logger, $store, $form, $delivery, $order);
 
                     $entityManager->flush();
 
@@ -409,30 +411,32 @@ trait StoreTrait
         return $pricingManager->duplicateOrder($store, $fromOrder);
     }
 
-    public function subscriptionAction($storeId, $subscriptionId,
+    public function recurrenceRuleAction($storeId,
+        $recurrenceRuleId,
         Request $request,
+        DeliveryManager $deliveryManager,
         PricingManager $pricingManager,
         EntityManagerInterface $entityManager,
         LoggerInterface $logger,
     )
     {
-        $subscription = $entityManager
+        $recurrenceRule = $entityManager
             ->getRepository(RecurrenceRule::class)
-            ->find($subscriptionId);
+            ->find($recurrenceRuleId);
 
-        $store = $subscription->getStore();
+        $store = $recurrenceRule->getStore();
 
         // Currently the route is only accessible by ROLE_DISPATCHER,
         // so this check is not doing much, but it would be useful
         // if we decide to open the route to store owners
         $this->denyAccessUnlessGranted('view', $store);
 
-        $tempDelivery = $pricingManager->createDeliveryFromSubscription($subscription, Carbon::now()->format('Y-m-d'), false);
+        $tempDelivery = $deliveryManager->createDeliveryFromRecurrenceRule($recurrenceRule, Carbon::now()->format('Y-m-d'), false);
 
         $routes = $request->attributes->get('routes');
 
         $arbitraryPrice = null;
-        if ($arbitraryPriceTemplate = $subscription->getArbitraryPriceTemplate()) {
+        if ($arbitraryPriceTemplate = $recurrenceRule->getArbitraryPriceTemplate()) {
             $arbitraryPrice = new ArbitraryPrice($arbitraryPriceTemplate['variantName'], $arbitraryPriceTemplate['variantPrice']);
         }
 
@@ -446,22 +450,23 @@ trait StoreTrait
             $tempDelivery = $form->getData();
 
             $arbitraryPrice = $this->getArbitraryPrice($form);
-            $rule = $this->getRecurrenceRule($form, $logger);
+            $recurrRule = $this->getRecurrRule($form, $logger);
 
-            if (null !== $rule) {
-                $pricingManager->updateSubscription($subscription, $tempDelivery, $rule, $arbitraryPrice ? new UseArbitraryPrice($arbitraryPrice) : new UsePricingRules);
+            if (null !== $recurrRule) {
+                $pricingManager->updateRecurrenceRule($recurrenceRule, $tempDelivery, $recurrRule, $arbitraryPrice ? new UseArbitraryPrice($arbitraryPrice) : new UsePricingRules);
                 $this->handleRememberAddress($store, $form);
                 $entityManager->flush();
             } else {
-                $pricingManager->cancelSubscription($subscription, $tempDelivery);
+                $pricingManager->cancelRecurrenceRule($recurrenceRule, $tempDelivery);
             }
 
-            return $this->redirectToRoute($routes['success']);
+            $redirectUri = $form->has('__redirect_to') ? $form->get('__redirect_to')->getData() : null;
+            return $redirectUri ? $this->redirect($redirectUri) : $this->redirectToRoute($routes['redirect_default'], ['id' => $storeId]);
         }
 
         return $this->render('store/subscriptions/item.html.twig', [
             'layout' => $request->attributes->get('layout'),
-            'subscription' => $subscription,
+            'recurrenceRule' => $recurrenceRule,
             'delivery' => $tempDelivery,
             'form' => $form->createView(),
         ]);
@@ -496,7 +501,7 @@ trait StoreTrait
         $orderManager->setBookmark($order, $isBookmarked);
     }
 
-    private function handleNewSubscription(
+    private function handleNewRecurrenceRule(
         PricingManager $pricingManager,
         LoggerInterface $logger,
         Store $store,
@@ -509,19 +514,19 @@ trait StoreTrait
             return;
         }
 
-        $recurrenceRule = $this->getRecurrenceRule($form, $logger);
+        $recurrRule = $this->getRecurrRule($form, $logger);
 
-        if (null === $recurrenceRule) {
+        if (null === $recurrRule) {
             return;
         }
 
-        $subscription = $pricingManager->createSubscription($store, $delivery, $recurrenceRule, $pricingStrategy);
+        $recurrenceRule = $pricingManager->createRecurrenceRule($store, $delivery, $recurrRule, $pricingStrategy);
 
-        if (null !== $subscription) {
-            $order->setSubscription($subscription);
+        if (null !== $recurrenceRule) {
+            $order->setSubscription($recurrenceRule);
 
             foreach ($delivery->getTasks() as $task) {
-                $task->setRecurrenceRule($subscription);
+                $task->setRecurrenceRule($recurrenceRule);
             }
         }
     }
@@ -547,7 +552,7 @@ trait StoreTrait
     }
 
 
-    private function getRecurrenceRule(FormInterface $form, LoggerInterface $logger): ?Rule
+    private function getRecurrRule(FormInterface $form, LoggerInterface $logger): ?Rule
     {
         if (!$form->has('recurrence')) {
             return null;
@@ -599,7 +604,9 @@ trait StoreTrait
         MessageBusInterface $messageBus,
         EntityManagerInterface $entityManager,
         Hashids $hashids8,
-        Filesystem $deliveryImportsFilesystem)
+        Filesystem $deliveryImportsFilesystem,
+        SlugifyInterface $slugify,
+    )
     {
         $store = $this->getDoctrine()
             ->getRepository(Store::class)
@@ -622,7 +629,8 @@ trait StoreTrait
                 messageBus: $messageBus,
                 entityManager: $entityManager,
                 hashids: $hashids8,
-                filesystem: $deliveryImportsFilesystem
+                filesystem: $deliveryImportsFilesystem,
+                slugify: $slugify
             );
         }
 
@@ -697,7 +705,7 @@ trait StoreTrait
         ]);
     }
 
-    public function storeDeliveriesBookmarksAction($id, Request $request,
+    public function storeSavedOrdersAction($id, Request $request,
         EntityManagerInterface $entityManager,
         OrderRepository $orderRepository)
     {
@@ -715,13 +723,66 @@ trait StoreTrait
 
         $routes = $request->attributes->get('routes');
 
-        return $this->render('store/deliveries_bookmarks.html.twig', [
+        return $this->render('store/orders_saved.html.twig', [
             'layout' => $request->attributes->get('layout'),
             'store' => $store,
             'bookmarks' => $orderRepository->findBookmarked($store, $this->getUser()),
             'routes' => $this->getDeliveryRoutes(),
             'stores_route' => $routes['stores'],
             'store_route' => $routes['store'],
+            'store_addresses_route' => $routes['store_addresses'],
+        ]);
+    }
+
+    public function storeRecurrenceRulesAction($id, Request $request,
+        EntityManagerInterface $entityManager,
+        DeliveryManager $deliveryManager,
+        PricingManager $pricingManager
+    )
+    {
+        $store = $entityManager
+            ->getRepository(Store::class)
+            ->find($id);
+
+        $this->accessControl($store, 'view');
+
+        $routes = $request->attributes->get('routes');
+
+        $data = [];
+        $this->entityManager->getFilters()->enable('soft_deleteable');
+        $recurrenceRules = $this->entityManager->getRepository(RecurrenceRule::class)->findBy(
+            array('store' => $store),
+            array('createdAt' => 'DESC')
+        );
+
+        foreach ($recurrenceRules as $rule) {
+            $templateOrder = $pricingManager->createOrderFromRecurrenceRule($rule, Carbon::now()->format('Y-m-d'), false);
+            $templateDelivery = $templateOrder ? $templateOrder->getDelivery() : $deliveryManager->createDeliveryFromRecurrenceRule($rule, Carbon::now()->format('Y-m-d'), false);
+            $templateTasks = $templateDelivery ? $templateDelivery->getTasks() : $deliveryManager->createTasksFromRecurrenceRule($rule, Carbon::now()->format('Y-m-d'), false);
+
+            $isLegacy = (null === $templateDelivery); // consider rules created from Dispatch dashboard that cannot be used to create a valid delivery as legacy
+
+            $data[] = [
+                'recurrenceRule' => $rule,
+                'templateTasks' => $templateTasks,
+                'templateDelivery' => $templateDelivery, // might be null if we cannot create a valid delivery (could be the case for some recurrence rules created from Dispatch dashboard)
+                'templateOrder' => $templateOrder, // might be null if we cannot calculate the price
+                'generateOrders' => !$isLegacy && $rule->isGenerateOrders(), //in the future that will be configurable
+                'isLegacy' => $isLegacy,
+            ];
+        }
+        $this->entityManager->getFilters()->disable('soft_deleteable');
+
+        return $this->render('store/recurrence_rules.html.twig', [
+            'layout' => $request->attributes->get('layout'),
+            'store' => $store,
+            'recurrence_rules' => $data,
+            'routes' => [
+                'view' => $routes['store_recurrence_rule'],
+            ],
+            'stores_route' => $routes['stores'],
+            'store_route' => $routes['store'],
+            'store_addresses_route' => $routes['store_addresses'],
         ]);
     }
 
@@ -831,8 +892,11 @@ trait StoreTrait
         Hashids $hashids,
         Filesystem $filesystem,
         MessageBusInterface $messageBus,
+        SlugifyInterface $slugify,
         string $routeTo)
-    {
+    {   
+        
+        /** @var UploadedFile $uploadedFile */
         $uploadedFile = $form->get('file')->getData();
 
         $queue = new DeliveryImportQueue();
@@ -841,7 +905,17 @@ trait StoreTrait
         $entityManager->persist($queue);
         $entityManager->flush();
 
-        $filename = sprintf('%s.%s', $hashids->encode($queue->getId()), $uploadedFile->guessExtension());
+        // slugify to get a safe filename. it may return an empty string but we had a hash afterwards so it should be alright.
+        $filename = $slugify->slugify($uploadedFile->getClientOriginalName()); 
+        $filename = substr($filename, 0, -4); // remove .csv extension in filename
+        $filename = substr($filename, 0, 100); // limit the filename length
+
+        $filename = sprintf(
+            '%s_%s.%s',
+            $filename,
+            $hashids->encode($queue->getId()),
+            $uploadedFile->guessExtension()
+        );
 
         try {