+++ src/Entity/Delivery.php
@@ -9,6 +9,7 @@ use AppBundle\Action\Delivery\Cancel as CancelDelivery;
 use AppBundle\Action\Delivery\Create as CreateDelivery;
 use AppBundle\Action\Delivery\Drop as DropDelivery;
 use AppBundle\Action\Delivery\Pick as PickDelivery;
+use AppBundle\Action\Delivery\BulkAsync as BulkAsyncDelivery;
 use AppBundle\Action\Delivery\SuggestOptimizations as SuggestOptimizationsController;
 use AppBundle\Api\Dto\DeliveryInput;
 use AppBundle\Api\Dto\OptimizationSuggestions;
@@ -77,6 +78,14 @@ use Symfony\Component\Serializer\Annotation\Groups;
  *         "summary"="Suggests optimizations for a delivery",
  *         "parameters"=Delivery::OPENAPI_CONTEXT_POST_PARAMETERS
  *       }
+ *     },
+ *     "deliveries_import_async"={
+ *       "method"="POST",
+ *       "path"="/deliveries/import_async",
+ *       "deserialize"=false,
+ *       "input_formats"={"csv"={"text/csv"}},
+ *       "controller"= BulkAsyncDelivery::class,
+ *       "security"="is_granted('ROLE_OAUTH2_DELIVERIES')"
  *     },
  *   },
  *   itemOperations={
@@ -294,6 +303,25 @@ class Delivery extends TaskCollection implements TaskCollectionInterface, Packag
         return new self();
     }
 
+    public static function canCreateWithTasks(Task ...$tasks): bool
+    {
+        if (count($tasks) < 2) {
+            return false;
+        }
+
+        // the first task must be a pickup
+        if (!$tasks[0]->isPickup()) {
+            return false;
+        }
+
+        // the last task must be a dropoff
+        if (!$tasks[count($tasks) - 1]->isDropoff()) {
+            return false;
+        }
+
+        return true;
+    }
+
     public static function createWithTasks(Task ...$tasks)
     {
         $delivery = self::create();
@@ -389,6 +417,16 @@ class Delivery extends TaskCollection implements TaskCollectionInterface, Packag
     public function isAssigned()
     {
         return $this->getPickup()->isAssigned() && $this->getDropoff()->isAssigned();
+    }
+
+    public function assignTo(User $user): void
+    {
+        $tasks = $this->getTasks();
+        array_walk($tasks,
+            function (Task $task) use ($user) {
+                $task->assignTo($user);
+            }
+        );
     }
 
     public function isCompleted()