+++ src/Entity/TaskRepository.php
@@ -7,6 +7,8 @@ use AppBundle\Entity\User;
 use AppBundle\Entity\TaskCollectionItem;
 use AppBundle\Entity\TaskList;
 use DateTime;
+use AppBundle\Utils\Barcode\Barcode;
+use AppBundle\Utils\Barcode\BarcodeUtils;
 use Doctrine\ORM\EntityRepository;
 use Doctrine\ORM\QueryBuilder;
 use Doctrine\ORM\Query\Expr;
@@ -72,5 +74,27 @@ class TaskRepository extends EntityRepository
             ->setParameter('subscription', $subscription)
             ->getQuery()
             ->getResult();
+    }
+
+    public function findByBarcode(string $barcode): ?Task
+    {
+        $barcode = BarcodeUtils::parse($barcode);
+
+        if ($barcode->isInternal() && $barcode->getEntityType() === Barcode::TYPE_TASK) {
+            return $this->find($barcode->getEntityId());
+        }
+
+        /* TODO: I'm using raw sql here, but doctrine doesn't support json fields and i don't
+                 want to install doctrine json functions just for this
+            @see https://github.com/ScientaNL/DoctrineJsonFunctions
+        */
+        $conn = $this->getEntityManager()->getConnection();
+        $sql = "SELECT id FROM task WHERE metadata->>'barcode' = :barcode";
+        $stmt = $conn->prepare($sql);
+        $id = $stmt->executeQuery(['barcode' => $barcode->getRawBarcode()])->fetchOne();
+        if ($id) {
+            return $this->find($id);
+        }
+        return null;
     }
 }