+++ src/Twig/CoopCycleExtension.php
@@ -68,6 +68,7 @@ class CoopCycleExtension extends AbstractExtension
             new TwigFilter('parse_expression', array(ExpressionLanguageRuntime::class, 'parseExpression')),
             new TwigFilter('expand_tags', array(TagsRuntime::class, 'expandTags')),
             new TwigFilter('placeholder_image', array(AssetsRuntime::class, 'placeholderImage')),
+            new TwigFilter('recurr_rule', array(RecurrRuleFormatResolver::class, 'format'), ['needs_context' => true]),
         );
     }
 
@@ -105,6 +106,7 @@ class CoopCycleExtension extends AbstractExtension
             new TwigFunction('restaurant_menu', array(LocalBusinessRuntime::class, 'resolveMenu')),
             new TwigFunction('order_phone_number', array(LocalBusinessRuntime::class, 'resolvePhoneNumber')),
             new TwigFunction('restaurant_opening_hours', array(LocalBusinessRuntime::class, 'openingHours')),
+            new TwigFunction('add_title_prefix', array(PageTitlePrefixResolver::class, 'addTitlePrefix')),
         );
     }